Fix setup error
