+++ openedx/core/lib/xblock_builtin/__init__.py
@@ -3,7 +3,6 @@ Helper functions shared by built-in XBlocks.
 
 """
 
-import os
 
 from django.conf import settings
 
@@ -30,16 +29,3 @@ def get_js_dependencies(group):
         return [settings.PIPELINE_JS[group]['output_filename']]
     else:
         return settings.PIPELINE_JS[group]['source_filenames']
-
-
-def package_data(pkg, root_list):
-    """
-    Generic function to find package_data for `pkg` under `root`.
-    """
-    data = []
-    for root in root_list:
-        for dirname, _, files in os.walk(os.path.join(pkg, root)):
-            for fname in files:
-                data.append(os.path.relpath(os.path.join(dirname, fname), pkg))
-
-    return {pkg: data}
