+++ openedx/core/lib/xblock_builtin/xblock_video/setup.py
@@ -2,9 +2,21 @@
 Setup for the Video XBlock.
 """
 
+import os
 from setuptools import find_packages, setup
 
-from openedx.core.lib.xblock_builtin import package_data
+
+def package_data(pkg, root_list):
+    """
+    Generic function to find package_data for `pkg` under `root`.
+    """
+    data = []
+    for root in root_list:
+        for dirname, _, files in os.walk(os.path.join(pkg, root)):
+            for fname in files:
+                data.append(os.path.relpath(os.path.join(dirname, fname), pkg))
+
+    return {pkg: data}
 
 
 setup(
